+++ README.md
@@ -1,15 +1,15 @@
-# Graphene
+# Inspector Graphetto
 
-_Graphene_ is a Gradle build plugin to visualize the task execution graph of your build. 
+_Inspector Graphetto_ is a Gradle build plugin to inspect and visualize the task execution graph of your build. 
 
 ## Installation
 
-Inside your top-level Gradle file add the Graphene plugin:
+Inside your top-level Gradle file add the Inspector Graphetto plugin:
 
 **In a `build.gradle` file (Groovy):**
 ```groovy
 plugins {
-    id 'at.droiddave.graphene' version '1.0.0-SNAPSHOT'
+    id 'at.droiddave.graphetto' version '1.0.0-SNAPSHOT'
 }
 ```
 
@@ -17,13 +17,13 @@ plugins {
 
 ```kotlin 
 plugins {
-    id("at.droiddave.graphene") version "1.0.0-SNAPSHOT"
+    id("at.droiddave.graphetto") version "1.0.0-SNAPSHOT"
 }
 ```
 
 ## Usage
 
-Simply run a build, and Graphene will write a GraphViz `.dot` file to `build/reports/taskGraph/graph.dot`.
+Simply run a build, and Inspector Graphetto will write a GraphViz `.dot` file to `build/reports/taskGraph/graph.dot`.
 
 ```shell
 $ ./gradlew assembleDebug
@@ -39,7 +39,7 @@ strict digraph G {
 You can also print the entire task dependency graph to the console:
 
 ```shell 
-$ ./gradlew assembleDebug -Dat.droiddave.graphene.consoleOutput=TREE
+$ ./gradlew assembleDebug -Dat.droiddave.graphetto.consoleOutput=TREE
 
  ── :someTask
     └── :someOtherTask
@@ -53,12 +53,12 @@ BUILD SUCCESSFUL in 92ms
 
 ## Configuration
 
-The plugin registers an extension called `graphene` on your project which can be used to configure the output:
+The plugin registers an extension called `graphetto` on your project which can be used to configure the output:
 
 ```groovy
-graphene {
+graphetto {
     outputFile = new File(buildDir, 'reports/my-task-graph.dot')
-    consoleOutput = at.droiddave.graphene.ConsoleOutput.TREE
+    consoleOutput = at.droiddave.graphetto.ConsoleOutput.TREE
 }
 ```
 
@@ -67,7 +67,7 @@ graphene {
 Configures the path of the `.dot` output file containing the information about the task graph that was executed. Defaults to `reports/taskGraph/graph.dot`.  
 
 ```groovy
-graphene {
+graphetto {
     outputFile = new File(buildDir, 'reports/my-task-graph.dot')
 }
 ```
@@ -80,9 +80,9 @@ Configures the type of console output produced by the plugin.
 * `NONE` will not print any console output. This is the default.
 
 ```groovy
-graphene {
-    consoleOutput = at.droiddave.graphene.ConsoleOutput.TREE
+graphetto {
+    consoleOutput = at.droiddave.graphetto.ConsoleOutput.TREE
 }
 ```
 
-You can also override the current configuration by passing the `-Dat.droiddave.graphene.consoleOutput=TREE` option when invoking your build on the command line.
+You can also override the current configuration by passing the `-Dat.droiddave.graphetto.consoleOutput=TREE` option when invoking your build on the command line.
