+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = 'graphene'
+rootProject.name = 'graphetto'
