+++ src/test/resources/fixtures/android-agp-3.5.2/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest
-        package="at.droiddave.graphene.testapp"
+        package="at.droiddave.inspectorgraphetto.testapp"
 />
